add bottom sheet list
