+++ DDUBUCK/app/src/main/res/layout/sheet_item.xml
@@ -10,8 +10,7 @@
     android:layout_marginBottom="8dp"
     android:paddingTop="16dp"
     android:paddingStart="16dp"
-    android:paddingEnd="0dp"
-    android:layout_width="360dp"
+    android:layout_width="match_parent"
     android:layout_height="match_parent">
 
     <ImageView
